+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.2.0 In progress
   * Added stripes-kint-components to dev deps
   * Tweaked Harnesses so that we can export *just* the base harness without the kint intl stuff. This means that kint-components can use the base harness as well.
+  * ERM-2773 Add agreement line with internal KB resource
+    * add packages: uuid, cypress-wait-until 
 
 ## 1.1.0 2023-02-22
   * Initial e2e cypress tests